refactor: split schemas.py to note.py and category.py
